+++ packages/terra-cell-grid/tests/wdio/cell-grid-spec.js
@@ -1,0 +1,16 @@
+Terra.describeViewports('CellGrid', ['medium'], () => {
+  describe('Cell Test', () => {
+    before(() => browser.url('/#/raw/tests/terra-cell-grid/cell-grid/cell'));
+    Terra.it.validatesElement();
+  });
+
+  describe('CellGrid Test', () => {
+    before(() => browser.url('/#/raw/tests/terra-cell-grid/cell-grid/cell-grid'));
+    Terra.it.validatesElement();
+  });
+
+  describe('CellGridDivider Test', () => {
+    before(() => browser.url('/#/raw/tests/terra-cell-grid/cell-grid/cell-grid-divider'));
+    Terra.it.validatesElement();
+  });
+});