+++ packages/terra-table/CHANGELOG.md
@@ -9,6 +9,13 @@ Unreleased
 ### Removed
 * Removed yarn reference in docs
 
+### Breaking Changes
+* terra-table now only provides controlled variants.
+* Removed Component `SingleSelectTableRows` - No longer supported this uncontrolled variant of table.
+* Removed Component `MultiSelectTableRows` - No longer supported this uncontrolled variant of table.
+* Removed Component `SelectableTableRows` - Selection by index is no longer provided.
+* Removed Component `TableRows` - body element now provided by default within the table.
+
 3.29.0 - (December 3, 2019)
 ------------------
 ### Added