+++ packages/terra-table/CHANGELOG.md
@@ -6,6 +6,15 @@ Unreleased
 ### Added
 * Added clinical-lowlight-theme and generated the respective reference wdio screenshots.
 
+4.0.0 - (January 17, 2020)
+------------------
+### Breaking Changes
+* terra-table now only provides controlled variants.
+* Removed Component `SingleSelectTableRows` - No longer supported this uncontrolled variant of table.
+* Removed Component `MultiSelectTableRows` - No longer supported this uncontrolled variant of table.
+* Removed Component `SelectableTableRows` - Selection by index is no longer provided.
+* Removed Component `TableRows` - body element now provided by default within the table.
+
 3.30.0 - (January 7, 2020)
 ------------------
 ### Removed